added ps support

+++ xv6-public/proc.h
@@ -35,10 +35,10 @@ struct context {
 enum procstate { UNUSED, EMBRYO, SLEEPING, RUNNABLE, RUNNING, ZOMBIE };
 
 struct cfsstate {
-  int nice;                         // Niceness of process (range 0-39, default 20)
-  int timeslice;                    // Timeslice of process [mticks] (max is 682 so int is fine here)
-  long long unsigned int runtime;   // Total runtime of process [mticks]
-  long long unsigned int vruntime;  // Total virtual runtime of process [mticks]
+  int nice;                   // Niceness of process (range 0-39, default 20)
+  int timeslice;              // Timeslice of process [mticks] (max is 682 so int is fine here)
+  unsigned int runtime;       // Total runtime of process [mticks]
+  unsigned int vruntime;      // Total virtual runtime of process [mticks]
 };
 
 // Per-process state
